REPO-2979: fixed incorrect test logic while investigating this issue.
I don't believe that the fix for MNT-18546 caused the two test run failures
(logged against REPO-2979) - indeed the test run immediately after the fix
was committed actually passed.

While I was investigating this bug report, I noticed that the test code
in TestCMIS.checkSecondaryTypes has a latent bug in it. We've been ok since
it doesn't exhibit problems when called with expectedMissingSecondaryTypes
having a size of 1, but would fail to pick up a situation where:

   *not all* the aspects that should have been removed were removed.

It would only have picked up the situation where:

   *none* of the aspects that should have been removed were removed.